Follow up fixes and improvements to new directions (#157)
> Follow up to #154 and #156 

I identified mutlitple issues in layout management were due to the fact
that `hexx` neighbors and directions were following a counter clockwise
order. This was causing some problems in angle management and visual
errors on pointy layouts.

Therefore, breaking change: From now on the hexx neighbors and
directions will follow the same order of the redblobgames docs, a
clockwise order.

I also added a new example, `hex_area` to reproduce one of the issues
found in #154 which also showcases how to use two different layouts in a
same grid.